refactor(login-screen): change Login Screen location
moved from screens folder to screens/app folder
