+++ pages/app/login.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginScreen from '../../src/components/screens/LoginScreen';
+import LoginScreen from '../../src/components/screens/app/LoginScreen';
 import websitePageHOC from '../../src/components/wrappers/WebsitePage/hoc';
 
 function LoginPage() {
